% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyTempdata.R
\docType{data}
\name{dummyTempdata}
\alias{dummyTempdata}
\title{Temperature data from place x}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{\code{id}}{integer id of weather station}
\item{\code{temp}}{double temperature in °C}
}
}
\usage{
dummyTempdata
}
\description{
Temperature measurements (2m height) measured on March 21st, 2009
}
\details{
DETAILS
}
\keyword{datasets}
